+++ src/pipecat/services/google.py
@@ -40,14 +40,10 @@ class GoogleLLMService(LLMService):
     franca for all LLM services, so that it is easy to switch between different LLMs.
     """
 
-    def __init__(self, model="gemini-1.5-flash-latest", api_key=None, **kwargs):
+    def __init__(self, api_key: str, model: str = "gemini-1.5-flash-latest", **kwargs):
         super().__init__(**kwargs)
-        self.model = model
-        gai.configure(api_key=api_key or os.environ["GOOGLE_API_KEY"])
-        self.create_client()
-
-    def create_client(self):
-        self._client = gai.GenerativeModel(self.model)
+        gai.configure(api_key=api_key)
+        self._client = gai.GenerativeModel(model)
 
     def _get_messages_from_openai_context(
             self, context: OpenAILLMContext) -> List[glm.Content]: