+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to **pipecat** will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- GoogleLLMService `api_key` argument is now mandatory.
+
+### Fixed
+
+- Fixed AzureLLMService.
+
 ## [0.0.23] - 2024-05-23
 
 ### Fixed